# Remove duplicate RSA call from AmazonCloudFrontUrlSigner