--- a/generator/.DevConfigs/36b119c0-a4f6-4944-81d0-856a037b43e4.json
+++ b/generator/.DevConfigs/36b119c0-a4f6-4944-81d0-856a037b43e4.json
@@ -1,0 +1,11 @@
+{
+  "services": [
+    {
+      "serviceName": "CloudFront",
+      "type": "patch",
+      "changeLogMessages": [
+        "Remove duplicate `RSA.ImportParameters` call from AmazonCloudFrontUrlSigner (which caused a performance regression issue in OpenSSL 3.0)"
+      ]
+    }
+  ]
+}