## Description
Our `AmazonCloudFrontUrlSigner` helper class invokes the `RSACryptoServiceProvider.ImportParameters` method twice, but it ignores the results of the first call. Customers using OpenSSL 3 have reported a performance regression as that operation is slower than previous versions of OpenSSL (see #3165 and #3172).

## Testing
- Dry-run: `DRY_RUN-4b9e956b-2815-4fc2-9d55-656d5938c53d`
- With the dry-run artifacts, I ran benchmark tests on an Ubuntu 2022 EC2 instance (benchmarks are based on the example from #3172):

```
$ uname -a
Linux ip-172-31-37-170 6.2.0-1018-aws #18~22.04.1-Ubuntu SMP Wed Jan 10 22:54:16 UTC 2024 x86_64 x86_64 x86_64 GNU/Linux

$ openssl version
OpenSSL 3.0.2 15 Mar 2022 (Library: OpenSSL 3.0.2 15 Mar 2022)

$ curl -H "X-aws-ec2-metadata-token: $TOKEN" http://169.254.169.254/latest/meta-data/instance-type
m7i.large
```

Before:
```
BenchmarkDotNet v0.13.12, Ubuntu 22.04.3 LTS (Jammy Jellyfish)
Intel Xeon Platinum 8488C, 1 CPU, 2 logical cores and 1 physical core
.NET SDK 8.0.101
  [Host]     : .NET 8.0.1 (8.0.123.58001), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 8.0.1 (8.0.123.58001), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
```
| Method                      | Mean     | Error    | StdDev   | Allocated |
|---------------------------- |---------:|---------:|---------:|----------:|
| CloudFrontGenerateSignedUrl | 63.64 ms | 0.102 ms | 0.086 ms |  37.29 KB |
| WildCardCloudFrontSignedUrl | 63.62 ms | 0.093 ms | 0.077 ms |  38.84 KB |

After:
```
BenchmarkDotNet v0.13.12, Ubuntu 22.04.3 LTS (Jammy Jellyfish)
Intel Xeon Platinum 8488C, 1 CPU, 2 logical cores and 1 physical core
.NET SDK 8.0.101
  [Host]     : .NET 8.0.1 (8.0.123.58001), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 8.0.1 (8.0.123.58001), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
```
| Method                      | Mean     | Error    | StdDev   | Allocated |
|---------------------------- |---------:|---------:|---------:|----------:|
| CloudFrontGenerateSignedUrl | 32.10 ms | 0.092 ms | 0.077 ms |  30.53 KB |
| WildCardCloudFrontSignedUrl | 32.11 ms | 0.081 ms | 0.072 ms |  32.02 KB |


## Types of changes
- [X] Bug fix (non-breaking change which fixes an issue)

## Checklist
- [X] My code follows the code style of this project
- [X] I have read the **README** document
- [X] I have added tests to cover my changes
- [X] All new and existing tests passed

## License
- [X] I confirm that this pull request can be released under the Apache 2 license