@SonalCalcey 2024-02-08
@dscpinheiro As for the above benchmarks the performance went from 60ms to 30ms that's good. but these [#3172](https://github.com/aws/aws-sdk-net/issues/3172) benchmark shows it is only taking 5ms in Ubuntu 20 version right?