@dscpinheiro 2024-02-08
> @dscpinheiro As for the above benchmarks the performance went from 60ms to 30ms that's good. but these [#3172](https://github.com/aws/aws-sdk-net/issues/3172) benchmark shows it is only taking 5ms in Ubuntu 20 version right?

I saw similar numbers on Ubuntu 20 (5ms before and 3.5ms after removing this duplicate call), that operation is just unfortunately slower in OpenSSL 3.

I know it doesn’t help your .NET 7 use case, but apparently there’ll be improvements included in .NET 8.0.3: https://github.com/dotnet/runtime/issues/97727#issuecomment-1930809385