@dscpinheiro 2024-02-08
This PR was included in version `3.7.301.41` of the `AWSSDK.CloudFront` package. I'm afraid the operation will still be slower when using OpenSSL 3 (compared to OpenSSL 1), but now the AWS SDK won't make the expensive call twice (and hopefully we'll see improvements in .NET 8.0.3 as well).

And as I mentioned above, we do have a separate issue for improving the overall performance of the CloudFront signers: https://github.com/aws/aws-sdk-net/issues/3130